perf(encode): stream finished 0.5 spans straight to msgpack
Extend the raw-span streaming path to the v0.5 agent wire. The shared `ByteSink`
serializes meta/metrics through the encoder's virtual `_encodeString`, which the
v0.5 encoder emits as string-table indices, so 0.5 reuses it unchanged; only the
per-span head differs (a 12-element index array vs the v0.4 map), which moves
into an overridable `_emitRawSpan`. Saves 192 ns/span (27.9%) on a representative
Express root span.

v0.5 emits strings as indices into a per-payload table, and the streaming walk
fills that table in a different order than the object path, so the two payloads
carry the same strings at different indices. The wire bytes therefore differ but
decode identically; the gate resolves both through their tables and compares the
decoded spans.
